Create WK4 Homework Create Math Quiz
Homework for creating math quiz in C#.

It gets two random numbers between 1 and 12 and multiplies them together. Asks the user to provide an int answer to the system and checks whether the answer is correct. 

The program keeps track of correct answers when achieved and prints results at the end.